Add lollipop router
